extend functionality for container start (check for "readystring" in log) add string for mssql db container
